refactor: 💡 unit-conversion service
se modifico la forma en que se maneja el stock en los insumos

+++ src/consumptions/entities/supplies-consumption-details.entity.ts
@@ -59,7 +59,7 @@ export class SuppliesConsumptionDetails {
   unit_of_measure: UnitType;
 
   @Column({
-    type: 'int4',
+    type: 'float8',
   })
   amount: number;
